requirements.txt
Flask>=3.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
reportlab>=4.0.0
matplotlib>=3.8.0
Faker>=24.0.0
python-dotenv>=1.0.1
    
